renamed paths and methods
