+++ src/main/java/com/dev/cinema/controllers/MovieController.java
@@ -1,9 +1,9 @@
 package com.dev.cinema.controllers;
 
-import com.dev.cinema.controllers.mapper.MovieMapper;
 import com.dev.cinema.model.dto.MovieRequestDto;
 import com.dev.cinema.model.dto.MovieResponseDto;
 import com.dev.cinema.service.MovieService;
+import com.dev.cinema.service.mapper.MovieMapper;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.web.bind.annotation.GetMapping;
