+++ src/main/java/com/dev/cinema/controllers/MovieSessionController.java
@@ -1,9 +1,9 @@
 package com.dev.cinema.controllers;
 
-import com.dev.cinema.controllers.mapper.MovieSessionMapper;
 import com.dev.cinema.model.dto.MovieSessionRequestDto;
 import com.dev.cinema.model.dto.MovieSessionResponseDto;
 import com.dev.cinema.service.MovieSessionService;
+import com.dev.cinema.service.mapper.MovieSessionMapper;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.stream.Collectors;
