+++ src/main/java/com/dev/cinema/controllers/OrderController.java
@@ -1,25 +1,23 @@
 package com.dev.cinema.controllers;
 
-import com.dev.cinema.controllers.mapper.OrderMapper;
 import com.dev.cinema.model.ShoppingCart;
 import com.dev.cinema.model.Ticket;
 import com.dev.cinema.model.User;
-import com.dev.cinema.model.dto.OrderRequestDto;
 import com.dev.cinema.model.dto.OrderResponseDto;
 import com.dev.cinema.service.OrderService;
 import com.dev.cinema.service.ShoppingCartService;
 import com.dev.cinema.service.UserService;
+import com.dev.cinema.service.mapper.OrderMapper;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/order")
+@RequestMapping("/orders")
 public class OrderController {
     private final OrderService orderService;
     private final OrderMapper orderMapper;
@@ -35,15 +33,15 @@ public class OrderController {
     }
 
     @PostMapping("/complete")
-    public void complete(@RequestBody OrderRequestDto orderDto) {
-        User user = userService.getById(orderDto.getUserId());
+    public void complete(@RequestParam Long userId) {
+        User user = userService.getById(userId);
         ShoppingCart cart = cartService.getByUser(user);
-        List<Ticket> ticket = cart.getTickets();
-        orderService.completeOrder(ticket, user);
+        List<Ticket> tickets = cart.getTickets();
+        orderService.completeOrder(tickets, user);
     }
 
-    @GetMapping("/get-order")
-    public List<OrderResponseDto> getOrderById(@RequestParam Long userId) {
+    @GetMapping
+    public List<OrderResponseDto> getOrderHistoryOfUser(@RequestParam Long userId) {
         return orderService.getOrderHistory(userService.getById(userId))
                 .stream().map(orderMapper::convertToResponseDto)
                 .collect(Collectors.toList());
