+++ src/main/java/com/dev/cinema/controllers/ShoppingCartController.java
@@ -1,6 +1,5 @@
 package com.dev.cinema.controllers;
 
-import com.dev.cinema.controllers.mapper.ShoppingCartMapper;
 import com.dev.cinema.model.MovieSession;
 import com.dev.cinema.model.ShoppingCart;
 import com.dev.cinema.model.User;
@@ -8,6 +7,7 @@ import com.dev.cinema.model.dto.ShoppingCartResponseDto;
 import com.dev.cinema.service.MovieSessionService;
 import com.dev.cinema.service.ShoppingCartService;
 import com.dev.cinema.service.UserService;
+import com.dev.cinema.service.mapper.ShoppingCartMapper;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/cart")
+@RequestMapping("/shopping-carts")
 public class ShoppingCartController {
     private final ShoppingCartMapper mapper;
     private final ShoppingCartService cartService;
@@ -30,7 +30,7 @@ public class ShoppingCartController {
         this.sessionService = sessionService;
     }
 
-    @GetMapping("/get")
+    @GetMapping("/by-user")
     public ShoppingCartResponseDto getUserById(@RequestParam Long id) {
         ShoppingCart cart = cartService.getByUser(userService.getById(id));
         return mapper.convertToResponseDto(cart);
