+++ src/main/java/com/dev/cinema/controllers/UserController.java
@@ -1,16 +1,16 @@
 package com.dev.cinema.controllers;
 
-import com.dev.cinema.controllers.mapper.UserMapper;
 import com.dev.cinema.model.User;
 import com.dev.cinema.model.dto.UserResponseDto;
 import com.dev.cinema.service.UserService;
+import com.dev.cinema.service.mapper.UserMapper;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/user")
+@RequestMapping("/users")
 public class UserController {
     private final UserService userService;
     private final UserMapper userMapper;
