--- src/main/java/com/dev/cinema/model/dto/OrderRequestDto.java
@@ -1,13 +1,0 @@
-package com.dev.cinema.model.dto;
-
-public class OrderRequestDto {
-    private Long userId;
-
-    public Long getUserId() {
-        return userId;
-    }
-
-    public void setUserId(Long userId) {
-        this.userId = userId;
-    }
-}
