+++ src/main/java/com/dev/cinema/model/dto/ShoppingCartResponseDto.java
@@ -4,24 +4,15 @@ import java.util.List;
 
 public class ShoppingCartResponseDto {
 
-    private Long id;
-    private List<Long> ticketsId;
+    private List<Long> ticketIds;
     private Long userId;
 
-    public Long getId() {
-        return id;
+    public List<Long> getTicketIds() {
+        return ticketIds;
     }
 
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public List<Long> getTicketsId() {
-        return ticketsId;
-    }
-
-    public void setTicketsId(List<Long> ticketsId) {
-        this.ticketsId = ticketsId;
+    public void setTicketIds(List<Long> ticketIds) {
+        this.ticketIds = ticketIds;
     }
 
     public Long getUserId() {
