+++ src/main/java/com/dev/cinema/model/dto/UserResponseDto.java
@@ -3,7 +3,6 @@ package com.dev.cinema.model.dto;
 public class UserResponseDto {
     private Long id;
     private String email;
-    private String password;
 
     public Long getId() {
         return id;
@@ -19,13 +18,5 @@ public class UserResponseDto {
 
     public void setEmail(String email) {
         this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
     }
 }
